Download blob before opening file for write
Creating the dst file before downloading the blob would yield an empty
file if the download should fail. E.g. if someone forgets to update the
sas-token. The empty file is problematic due to the isfile guard, as it
would make it impossible (without changing the code) to retry the
download.

Although easy to work-around, it can be a bit confussing at first and is
generally a bad design. Downloading the blob before opening a file for
writing is a more robust solution.