Ran notebook with figure DPI set to 50.
This lowers the notebook file size and should greatly increase the chances that it will be rendered when opened on github. The DPI is still high enough to ensure that figures are still quite readable. The notebook is still set to run itself at 100 DPI, so when it's rerun locally, it will produce much higher-quality figures.